+++ Sources/Web3Inbox/WebView/ChatWebViewEvent.swift
@@ -1,0 +1,14 @@
+import Foundation
+
+enum ChatWebViewEvent: String {
+    case getReceivedInvites
+    case getSentInvites
+    case getThreads
+    case register
+    case resolve
+    case getMessages
+    case message
+    case accept
+    case reject
+    case invite
+}
