+++ Sources/Web3Inbox/WebView/PushWebViewEvent.swift
@@ -1,0 +1,11 @@
+import Foundation
+
+enum PushWebViewEvent: String {
+    case approve
+    case reject
+    case subscribe
+    case getActiveSubscriptions
+    case getMessageHistory
+    case deleteSubscription
+    case deletePushMessage
+}
